+++ file_helper.h
@@ -12,8 +12,8 @@ void * file_open_readonly(const char *path);
 void * file_open_readwrite(const char *path);
 void * file_open_append(const char *path);
 
-int file_read(void *file, unsigned long long offset, unsigned char *data, unsigned int size);
-int file_write(void *file, unsigned long long offset, unsigned char *data, unsigned int size);
+int file_read(void *file, unsigned long long offset, unsigned char *data, size_t size);
+int file_write(void *file, unsigned long long offset, unsigned char *data, size_t size);
 
 long file_mode(const char * file);
 long file_length(const char * file);
